+++ backend/routes/api/card_management.py
@@ -71,6 +71,17 @@ RENAME_FLASHCARD_FORMAT = {
     "newName": ""
 }
 
+RENAME_FOLDER_FORMAT = {
+    "userID": "",
+    "currentName": "",
+    "newName": ""
+}
+
+DELETE_FOLDER_FORMAT = {
+    "userID": "",
+    "folder": ""
+}
+
 @card_management_routes.route("/api/create-flashcard", methods=["POST"])
 @validate_json(CREATE_FLASHCARD_FORMAT)
 def create_flashcard():
@@ -315,6 +326,7 @@ def update_card_progress():
         card_data = request.json.get("cardData")
 
         db.folders.update_card_progress(user_id, card_data)
+        db.statistics.increase_xp(user_id, len(card_data) * 10)
 
         return jsonify({"success": "Card progress updated"}), 200
     except Exception as e:
@@ -368,5 +380,55 @@ def rename_flashcard():
             return jsonify({"success": f"Flashcard {flashcard_id} renamed to {new_name}"}), 200
         else:
             return jsonify({"error": f"Flashcard {flashcard_id} does not exist"}), 404
+    except Exception as e:
+        return jsonify(str(e)), 500
+
+@card_management_routes.route("/api/rename-folder", methods=["POST"])
+@validate_json(RENAME_FOLDER_FORMAT)
+def rename_folder():
+    """
+    Rename a  folder
+    Example request:
+    {
+        "userID": "my-id",
+        "currentName": "my-folder",
+        "newName": "new-name"
+    }
+    """
+    try:
+        user_id = request.json.get("userID")
+        folder_name = request.json.get("currentName")
+        new_name = request.json.get("newName")
+
+        result = db.folders.rename_folder(user_id, folder_name, new_name)
+
+        if result is not None:
+            return jsonify({"success": f"Flashcard {folder_name} renamed to {new_name}"}), 200
+        else:
+            return jsonify({"error": f"Flashcard {folder_name} does not exist"}), 404
+    except Exception as e:
+        return jsonify(str(e)), 500
+
+@card_management_routes.route("/api/delete-folder", methods=["DELETE"])
+@validate_json(DELETE_FOLDER_FORMAT)
+def delete_folder():
+    """
+    Delete a folder
+    Example request:
+    {
+        "userID": "my-id",
+        "folder": "folder-name"
+    }
+    """
+    try:
+        user_id = request.json.get("userID")
+        folder_name = request.json.get("folder")
+
+        result = db.folders.delete_folder(user_id, folder_name)
+
+        if result is not None:
+            return jsonify({"success": f"Folder {folder_name} deleted"}), 200
+        else:
+            return jsonify({"error": f"Folder {folder_name} does not exist"}), 404
     except Exception as e:
         return jsonify(str(e)), 500