+++ frontend/src/App.js
@@ -5,8 +5,26 @@ import MainPage from "./screens/MainPage";
 import Flashcards from './screens/Flashcards';
 import EditFlashcard from './screens/EditFlashcard';
 import ViewFlashcards from './screens/ViewFlashcards';
+import PageNotFound from './screens/PageNotFound';
+import ErrorBoundary from './containers/ErrorBoundary/ErrorBoundary';
 import { getCookie } from './api/Authentication';
 
+function PromptLoginIfNotLoggedIn({ child, userID, setUserID }) {
+  return (
+    <>
+      {userID ? child : <SignInPage setUserID={setUserID}/>}
+    </>
+  )
+}
+
+function ErrorChecking({ userID, setUserID, child }) {
+  /* Add the ErrorBoundary and PromptLoginIfNotLoggedIn components */
+  return (
+    <ErrorBoundary>
+      <PromptLoginIfNotLoggedIn child={child} userID={userID} setUserID={setUserID}/>
+    </ErrorBoundary>
+  )
+}
 
 function App() {
   const [userID, setUserID] = useState(getCookie("userID"));
@@ -14,11 +32,12 @@ function App() {
     <>
       <BrowserRouter>
         <Routes>
-          <Route path="/" element={<SignInPage active={false}/>} />
-          <Route path="/dashboard" element={<MainPage userID={userID} setUserID={setUserID}/>} />
-          <Route path="/flashcards" element={<Flashcards />} />
-          <Route path="/edit-flashcard-set" element={<EditFlashcard />} />
-          <Route path="/view" element={<ViewFlashcards />} />
+          <Route path="/" element={<ErrorChecking userID={userID} setUserID={setUserID} child={<SignInPage active={false}/>} />} />
+          <Route path="/dashboard" element={<ErrorChecking userID={userID} setUserID={setUserID} child={<MainPage userID={userID} setUserID={setUserID}/>} />} />
+          <Route path="/flashcards" element={<ErrorChecking userID={userID} setUserID={setUserID} child={<Flashcards />} />} />
+          <Route path="/edit-flashcard-set" element={<ErrorChecking userID={userID} setUserID={setUserID} child={<EditFlashcard />} />} />
+          <Route path="/view" element={<ErrorChecking userID={userID} setUserID={setUserID} child={<ViewFlashcards />} />} />
+          <Route path="*" element={<ErrorChecking userID={userID} setUserID={setUserID} child={<PageNotFound />} />} />
         </Routes>
       </BrowserRouter>
     </>