+++ frontend/src/containers/FlashcardOverview/FlashcardOverview.js
@@ -2,12 +2,13 @@ import React from 'react';
 import FlashcardFolder from './FlashcardFolder/FlashcardFolder';
 import FlashcardItem from './FlashcardItem/FlashcardItem';
 import Heading5 from '../../componments/Text/Heading5/Heading5';
-
 function FlashcardOverview({
   flashcardData,
   setMoveFolderDialogueVisible,
-  showDeleteConfirmation,
+  showDeleteFlashcardConfirmation,
   setRenameFlashcardSetPopupVisible,
+  setDeleteFolderConfirmationVisible,
+  setRenameFolderPopupVisible,
   view,
   selected,
   setSelected
@@ -27,7 +28,7 @@ function FlashcardOverview({
           <FlashcardItem
             element={element}
             setMoveFolderDialogueVisible={setMoveFolderDialogueVisible}
-            showDeleteConfirmation={showDeleteConfirmation}
+            showDeleteConfirmation={showDeleteFlashcardConfirmation}
             setRenameFlashcardSetPopupVisible={setRenameFlashcardSetPopupVisible}
             flashcardData={filterFlashcardSet(flashcardData, element.flashcardName)}
             path={path}
@@ -49,6 +50,9 @@ function FlashcardOverview({
           element={element}
           name={folderName}
           folderKey={folderName + element.flashcardID}
+          showDeleteConfirmation={setDeleteFolderConfirmationVisible}
+          setRenameFolderPopupVisible={setRenameFolderPopupVisible}
+          flashcardData={element}
           child={Object.entries(element).map(([key, value]) => renderElement(value, key, path))}
           view={view}
           path={path}